ensure online cache is and can be refreshed
We use `COMPRESS_OFFLINE = True` which disable the online cache
generation. On Debian, the cache is not pre generated, unlike
Ubuntu.
The changes add be introduce in Ubuntu horizon 2012.2~rc1-0ubuntu1.

So, basicly, this changes ensure Horizon with the COMPRESS_OFFLINE=True
parameter will continue to work in this situations:
- Horizon template has been changes, python-lesscpy is required to
  regenerate the cache.
- the package doesn't provide pre-cached files (Debian), we generate the
  cache.

For the record, a issue has been opened on Debian bug tracker regard
this situation: http://bugs.debian.org/739698

See:
  I6671938a87744a4dd0b84021ca87cd49aea10741
  Ie290701b00c97352838e13182b940a6de46a591f

Change-Id: I813b5f6067bb6ecce279cab7278d9227c4d31d28